+++ R/create_image.R
@@ -1,71 +1,73 @@
-#' List fine-tune events
+#' Create image
 #'
-#' Returns events related to a specified fine-tune job. See [this
-#' page](https://beta.openai.com/docs/api-reference/fine-tunes/events) for
+#' Creates an image given a prompt. See [this
+#' page](https://beta.openai.com/docs/api-reference/images/create) for
 #' details.
 #'
 #' For arguments description please refer to the [official
-#' documentation](https://beta.openai.com/docs/api-reference/fine-tunes/events).
+#' documentation](https://beta.openai.com/docs/api-reference/images/create).
 #'
-#' @param fine_tune_id required; a length one character vector.
-#' @param stream required; defaults to `FALSE`; a length one logical vector.
-#'   **Currently is not implemented.**
+#' @param prompt required; a length one character vector.
+#' @param n required; defaults to `1`; a length one numeric vector with the
+#'   integer value greater than `0`.
+#' @param size required; defaults to `"1024x1024"`; a length one character
+#'   vector, one among `"256x256"`, `"512x512"`, and `"1024x1024"`.
+#' @param response_format required; defaults to `"url"`; a length one character
+#'   vector, one among `"url"` and `"b64_json"`.
+#' @param user optional; defaults to `NULL`; a length one character vector.
 #' @param openai_api_key required; defaults to `Sys.getenv("OPENAI_API_KEY")`
 #'   (i.e., the value is retrieved from the `.Renviron` file); a length one
 #'   character vector. Specifies OpenAI API key.
 #' @param openai_organization optional; defaults to `NULL`; a length one
 #'   character vector. Specifies OpenAI organization.
-#' @return Returns a list, elements of which contains information about the
-#'   fine-tune events.
+#' @return Returns a list, an element of which contain either a link to the
+#' generated image or the generated image decoded in Base64.
 #' @examples \dontrun{
-#' training_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#' validation_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#'
-#' training_info <- upload_file(training_file, "fine-tune")
-#' validation_info <- upload_file(validation_file, "fine-tune")
-#'
-#' info <- create_fine_tune(
-#'     training_file = training_info$id,
-#'     validation_file = validation_info$id,
-#'     model = "ada",
-#'     compute_classification_metrics = TRUE,
-#'     classification_positive_class = " baseball" # Mind space in front
-#' )
-#'
-#' id <- ifelse(
-#'     length(info$data$id) > 1,
-#'     info$data$id[length(info$data$id)],
-#'     info$data$id
-#' )
-#'
-#' list_fine_tune_events(fine_tune_id = id)
+#' create_image("An astronaut riding a horse in a photorealistic style")
 #' }
-#' @family fine-tune functions
+#' @family image functions
 #' @export
-list_fine_tune_events <- function(
-        fine_tune_id,
-        stream = FALSE,
+create_image <- function(
+        prompt,
+        n = 1,
+        size = c("1024x1024", "256x256", "512x512"),
+        response_format = c("url", "b64_json"),
+        user = NULL,
         openai_api_key = Sys.getenv("OPENAI_API_KEY"),
         openai_organization = NULL
 ) {
+
+    size <- match.arg(size)
+    response_format <- match.arg(response_format)
 
     #---------------------------------------------------------------------------
     # Validate arguments
 
     assertthat::assert_that(
-        assertthat::is.string(fine_tune_id),
-        assertthat::noNA(fine_tune_id)
+        assertthat::is.string(prompt),
+        assertthat::noNA(prompt)
     )
 
     assertthat::assert_that(
-        assertthat::is.flag(stream),
-        assertthat::noNA(stream),
-        is_false(stream)
+        assertthat::is.count(n)
     )
+
+    assertthat::assert_that(
+        assertthat::is.string(size),
+        assertthat::noNA(size)
+    )
+
+    assertthat::assert_that(
+        assertthat::is.string(response_format),
+        assertthat::noNA(response_format)
+    )
+
+    if (!is.null(user)) {
+        assertthat::assert_that(
+            assertthat::is.string(user),
+            assertthat::noNA(user)
+        )
+    }
 
     assertthat::assert_that(
         assertthat::is.string(openai_api_key),
@@ -80,11 +82,11 @@ list_fine_tune_events <- function(
     }
 
     #---------------------------------------------------------------------------
-    # Build parameters of the request
+    # Build path parameters
 
-    base_url <- glue::glue(
-        "https://api.openai.com/v1/fine-tunes/{fine_tune_id}/events"
-    )
+    task <- "images/generations"
+
+    base_url <- glue::glue("https://api.openai.com/v1/{task}")
 
     headers <- c(
         "Authorization" = paste("Bearer", openai_api_key),
@@ -99,12 +101,16 @@ list_fine_tune_events <- function(
     # Build request body
 
     body <- list()
-    body[["stream"]] <- stream
+    body[["prompt"]] <- prompt
+    body[["n"]] <- n
+    body[["size"]] <- size
+    body[["response_format"]] <- response_format
+    body[["user"]] <- user
 
     #---------------------------------------------------------------------------
     # Make a request and parse it
 
-    response <- httr::GET(
+    response <- httr::POST(
         url = base_url,
         httr::add_headers(.headers = headers),
         body = body,