+++ R/create_moderation.R
@@ -1,70 +1,46 @@
-#' List fine-tune events
+#' Create moderation
 #'
-#' Returns events related to a specified fine-tune job. See [this
-#' page](https://beta.openai.com/docs/api-reference/fine-tunes/events) for
+#' Classifies if text violates OpenAI's Content Policy. See [this
+#' page](https://beta.openai.com/docs/api-reference/moderations/create) for
 #' details.
 #'
 #' For arguments description please refer to the [official
-#' documentation](https://beta.openai.com/docs/api-reference/fine-tunes/events).
+#' documentation](https://beta.openai.com/docs/api-reference/completions/create).
 #'
-#' @param fine_tune_id required; a length one character vector.
-#' @param stream required; defaults to `FALSE`; a length one logical vector.
-#'   **Currently is not implemented.**
+#' @param input required; an arbitrary length character vector.
+#' @param model required; a length one character vector. One among
+#' `"text-moderation-stable"` and `"text-moderation-latest"`.
 #' @param openai_api_key required; defaults to `Sys.getenv("OPENAI_API_KEY")`
 #'   (i.e., the value is retrieved from the `.Renviron` file); a length one
 #'   character vector. Specifies OpenAI API key.
 #' @param openai_organization optional; defaults to `NULL`; a length one
 #'   character vector. Specifies OpenAI organization.
-#' @return Returns a list, elements of which contains information about the
-#'   fine-tune events.
+#' @return Returns a list, elements of which contain information about the
+#'   model.
 #' @examples \dontrun{
-#' training_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#' validation_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#'
-#' training_info <- upload_file(training_file, "fine-tune")
-#' validation_info <- upload_file(validation_file, "fine-tune")
-#'
-#' info <- create_fine_tune(
-#'     training_file = training_info$id,
-#'     validation_file = validation_info$id,
-#'     model = "ada",
-#'     compute_classification_metrics = TRUE,
-#'     classification_positive_class = " baseball" # Mind space in front
-#' )
-#'
-#' id <- ifelse(
-#'     length(info$data$id) > 1,
-#'     info$data$id[length(info$data$id)],
-#'     info$data$id
-#' )
-#'
-#' list_fine_tune_events(fine_tune_id = id)
+#' create_moderation("I want to kill them all.")
 #' }
-#' @family fine-tune functions
 #' @export
-list_fine_tune_events <- function(
-        fine_tune_id,
-        stream = FALSE,
+create_moderation <- function(
+        input,
+        model = c("text-moderation-stable", "text-moderation-latest"),
         openai_api_key = Sys.getenv("OPENAI_API_KEY"),
         openai_organization = NULL
 ) {
+
+    model <- match.arg(model)
 
     #---------------------------------------------------------------------------
     # Validate arguments
 
     assertthat::assert_that(
-        assertthat::is.string(fine_tune_id),
-        assertthat::noNA(fine_tune_id)
+        is.character(input),
+        assertthat::noNA(input)
     )
 
     assertthat::assert_that(
-        assertthat::is.flag(stream),
-        assertthat::noNA(stream),
-        is_false(stream)
+        assertthat::is.string(model),
+        assertthat::noNA(model)
     )
 
     assertthat::assert_that(
@@ -80,11 +56,11 @@ list_fine_tune_events <- function(
     }
 
     #---------------------------------------------------------------------------
-    # Build parameters of the request
+    # Build path parameters
 
-    base_url <- glue::glue(
-        "https://api.openai.com/v1/fine-tunes/{fine_tune_id}/events"
-    )
+    task <- "moderations"
+
+    base_url <- glue::glue("https://api.openai.com/v1/{task}")
 
     headers <- c(
         "Authorization" = paste("Bearer", openai_api_key),
@@ -99,12 +75,13 @@ list_fine_tune_events <- function(
     # Build request body
 
     body <- list()
-    body[["stream"]] <- stream
+    body[["input"]] <- input
+    body[["model"]] <- model
 
     #---------------------------------------------------------------------------
     # Make a request and parse it
 
-    response <- httr::GET(
+    response <- httr::POST(
         url = base_url,
         httr::add_headers(.headers = headers),
         body = body,