+++ R/list_models.R
@@ -1,71 +1,32 @@
-#' List fine-tune events
+#' List models
 #'
-#' Returns events related to a specified fine-tune job. See [this
-#' page](https://beta.openai.com/docs/api-reference/fine-tunes/events) for
+#' Lists the currently available models, and provides basic information about
+#' each one such as the owner and availability. See
+#' [this page](https://beta.openai.com/docs/api-reference/models/list) for
 #' details.
 #'
 #' For arguments description please refer to the [official
-#' documentation](https://beta.openai.com/docs/api-reference/fine-tunes/events).
+#' documentation](https://beta.openai.com/docs/api-reference/models/list).
 #'
-#' @param fine_tune_id required; a length one character vector.
-#' @param stream required; defaults to `FALSE`; a length one logical vector.
-#'   **Currently is not implemented.**
 #' @param openai_api_key required; defaults to `Sys.getenv("OPENAI_API_KEY")`
 #'   (i.e., the value is retrieved from the `.Renviron` file); a length one
 #'   character vector. Specifies OpenAI API key.
 #' @param openai_organization optional; defaults to `NULL`; a length one
 #'   character vector. Specifies OpenAI organization.
-#' @return Returns a list, elements of which contains information about the
-#'   fine-tune events.
+#' @return Returns a list, an element of which is a data frame containing
+#'   information about models.
 #' @examples \dontrun{
-#' training_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#' validation_file <- system.file(
-#'     "extdata", "sport_prepared_train.jsonl", package = "openai"
-#' )
-#'
-#' training_info <- upload_file(training_file, "fine-tune")
-#' validation_info <- upload_file(validation_file, "fine-tune")
-#'
-#' info <- create_fine_tune(
-#'     training_file = training_info$id,
-#'     validation_file = validation_info$id,
-#'     model = "ada",
-#'     compute_classification_metrics = TRUE,
-#'     classification_positive_class = " baseball" # Mind space in front
-#' )
-#'
-#' id <- ifelse(
-#'     length(info$data$id) > 1,
-#'     info$data$id[length(info$data$id)],
-#'     info$data$id
-#' )
-#'
-#' list_fine_tune_events(fine_tune_id = id)
+#' list_models()
 #' }
-#' @family fine-tune functions
+#' @family model functions
 #' @export
-list_fine_tune_events <- function(
-        fine_tune_id,
-        stream = FALSE,
+list_models <- function(
         openai_api_key = Sys.getenv("OPENAI_API_KEY"),
         openai_organization = NULL
 ) {
 
     #---------------------------------------------------------------------------
     # Validate arguments
-
-    assertthat::assert_that(
-        assertthat::is.string(fine_tune_id),
-        assertthat::noNA(fine_tune_id)
-    )
-
-    assertthat::assert_that(
-        assertthat::is.flag(stream),
-        assertthat::noNA(stream),
-        is_false(stream)
-    )
 
     assertthat::assert_that(
         assertthat::is.string(openai_api_key),
@@ -82,9 +43,7 @@ list_fine_tune_events <- function(
     #---------------------------------------------------------------------------
     # Build parameters of the request
 
-    base_url <- glue::glue(
-        "https://api.openai.com/v1/fine-tunes/{fine_tune_id}/events"
-    )
+    base_url <- "https://api.openai.com/v1/models"
 
     headers <- c(
         "Authorization" = paste("Bearer", openai_api_key),
@@ -96,18 +55,11 @@ list_fine_tune_events <- function(
     }
 
     #---------------------------------------------------------------------------
-    # Build request body
-
-    body <- list()
-    body[["stream"]] <- stream
-
-    #---------------------------------------------------------------------------
     # Make a request and parse it
 
     response <- httr::GET(
         url = base_url,
         httr::add_headers(.headers = headers),
-        body = body,
         encode = "json"
     )