+++ NEWS.md
@@ -1,3 +1,11 @@
+# openai 0.2.0
+
+* Add new DALL-E functions, namely `create_image()`, `create_image_edit()`, and `create_image_variation()`
+
+* Add the new function `create_moderation()` that checks whether the input violates OpenAI's Content Policy
+
+* Add new model-related functions, namely `list_models()` and `retrieve_model()`
+
 # openai 0.1.0
 
 * Initial version